no-mistakes(test): fix inverted canonical-twin payload restore guard in replay scan
